app/requirement.txt
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kombu==5.3.1
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
python-dateutil==2.8.2
redis==5.0.0
six==1.16.0
typing-extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
werkzeug==2.3.7
zipp==3.16.2
